requirements.txt
# --- Core Deep Learning & GNN ---
torch>=2.0.0
torch-geometric
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv

# --- Rough Path Theory (SigGNN) ---
signatory

# --- Tabular Experts (Boosting) ---
lightgbm
xgboost

# --- Data Engineering & Processing ---
pandas
numpy
scipy
scikit-learn
pyyaml
tqdm
joblib
pyarrow
fastparquet

# --- Experiment Tracking ---
wandb                  # Used in main.py for run logging and hyperparameter sweeps
torch-ema              # Used in trainer.py for EMA weight averaging (SOTA stability)

# --- Diagnostics & Visualization ---
matplotlib
seaborn
networkx
tensorboard